git-svn-id: https://svn.apache.org/repos/asf/incubator/kafka/trunk@1413738 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/scala/kafka/javaapi/message/MessageSet.scala
@@ -38,7 +38,7 @@ abstract class MessageSet extends java.lang.Iterable[MessageAndOffset] {
   /**
    * Gives the total size of this message set in bytes
    */
-  def sizeInBytes: Long
+  def sizeInBytes: Int
 
   /**
    * Validate the checksum of all the messages in the set. Throws an InvalidMessageException if the checksum doesn't
